brac/requirements.txt
gym
h5py
scipy
lxml==4.4.1
pyprind==2.11.2
nose2==0.8.0
six
path.py
joblib
python-dateutil
cached_property
cloudpickle
pyglet
matplotlib
imutils==0.5.1
numpydoc
ray
opencv-python
dill
lz4
setproctitle
psutil
opencv-python
boto3
redis
pandas
plotly
absl-py>=0.5.0
mujoco-py
numpy>=1.13.3
gin-config
tf-agents
